requirements.txt
line-bot-sdk==1.15.0
Flask==1.1.1
pya3rt==1.1
Keras==2.3.1
Pillow==7.0.0
pathlib==1.0.1
tensorflow==1.14.0
